refactor(llm): move model listing logic into the providers that own it
model_listing.ts had accreted three unrelated things over the feature's
seven commits. Only one of them was genuinely shared; the rest were
single-provider rules parked in a shared module for testability, which
the provider specs already cover through their mocked endpoints.

- OpenAI's chat-model filter, display-name derivation and newest-
  generation rule move into openai.ts; Gemini's chat-model filter into
  google.ts; the Claude per-family rule into anthropic.ts.
- recommendedModelIds becomes a method on LlmProvider, defaulting to the
  generic non-preview/non-legacy rule in BaseProvider. listProviderModels
  now calls instance.recommendedModelIds(models) instead of dispatching
  on a user-controlled provider string.
- claude-agent shares Anthropic's id shape, so it delegates to the
  helper exported from anthropic.ts rather than duplicating the regex —
  it implements LlmProvider directly and cannot inherit the override.
- mergeModelLists, RemoteModel and the price-table types have two
  independent consumers (BaseProvider and ClaudeAgentProvider) and no
  single owner, so they move to base_provider.ts, which claude_agent.ts
  already imported for buildModelList. update-model-prices.ts follows;
  its import stays type-only so the script pulls in no runtime deps.

The module's tests move with it, each into the spec of the provider that
now owns the rule.
