+++ apps/server/src/services/llm/providers/google.spec.ts
@@ -22,7 +22,7 @@ vi.mock("ai", async (importOriginal) => {
     return { ...actual, streamText: streamTextMock };
 });
 
-import { GoogleProvider } from "./google.js";
+import { GoogleProvider, isGoogleChatModel } from "./google.js";
 
 describe("GoogleProvider construction", () => {
     beforeEach(() => {
@@ -216,5 +216,62 @@ describe("GoogleProvider model listing", () => {
         fetchMock.mockResolvedValue(okJson({ models: { unexpected: "shape" } }));
         const provider = new GoogleProvider("g-key");
         await expect(provider.listModels()).rejects.toThrow(/Unexpected \/models response shape/);
+    });
+});
+
+describe("isGoogleChatModel", () => {
+    it("keeps Gemini chat models, including unknown future ones", () => {
+        for (const id of [
+            "gemini-2.5-pro",
+            "gemini-2.5-flash",
+            "gemini-2.5-flash-lite",
+            "gemini-3-flash-preview",
+            "gemini-3.1-pro-preview",
+            "gemini-3.5-flash",
+            "gemini-3.6-flash",
+            "gemini-2.0-flash-exp"
+        ]) {
+            expect(isGoogleChatModel(id), id).toBe(true);
+        }
+    });
+
+    it("drops non-Gemini families wholesale", () => {
+        for (const id of [
+            "text-embedding-004",
+            "aqa",
+            "imagen-3.0-generate-002",
+            "veo-3.1-generate-preview",
+            "lyria-3-pro-preview",
+            "gemma-4-26b-a4b-it",
+            "deep-research-preview-04-2026",
+            "antigravity-preview-05-2026",
+            "learnlm-2.0-flash-experimental"
+        ]) {
+            expect(isGoogleChatModel(id), id).toBe(false);
+        }
+    });
+
+    it("drops non-conversational gemini-* variants (media, robotics, computer use)", () => {
+        for (const id of [
+            "gemini-embedding-001",
+            "gemini-2.5-flash-image", // Nano Banana
+            "gemini-3-pro-image", // Nano Banana Pro
+            "gemini-3.1-flash-lite-image", // Nano Banana 2 Lite
+            "gemini-2.5-flash-preview-tts",
+            "gemini-3.1-flash-live-preview",
+            "gemini-2.5-flash-native-audio-preview-12-2025",
+            "gemini-omni-flash",
+            "gemini-robotics-er-1.6-preview",
+            "gemini-2.5-computer-use-preview-10-2025",
+            "gemini-3.1-pro-preview-customtools" // real id — no hyphen in "customtools"
+        ]) {
+            expect(isGoogleChatModel(id), id).toBe(false);
+        }
+    });
+
+    it("drops rolling aliases and pinned revisions that duplicate stable ids", () => {
+        for (const id of ["gemini-flash-latest", "gemini-pro-latest", "gemini-flash-lite-latest", "gemini-2.0-flash-lite-001"]) {
+            expect(isGoogleChatModel(id), id).toBe(false);
+        }
     });
 });
